stacks/apps/mealie-importer/app/requirements.txt
requests>=2.31.0
feedparser>=6.0.10
PyYAML>=6.0.1
beautifulsoup4>=4.12.0
urllib3>=2.0.0
lxml>=4.9.0
